requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
tqdm>=4.62.0

# PDF and Document Processing
PyPDF2>=2.0.0
python-docx>=0.8.11
pdfminer.six>=20201018
textract-msg>=0.28.7  # For .msg file support

# NLP and ML
nltk>=3.6.3
spacy>=3.2.0
scikit-learn>=1.0.2
sentence-transformers>=2.2.2

# Web Framework
streamlit>=1.9.0
streamlit-extras>=0.2.7

# Utilities
python-multipart>=0.0.5
